fix(export): preserve GlobalId as the IFCX node path in the Rust IFC5 exporter
ifc-lite export --format ifcx (and any other caller of export_ifc5) minted a
fresh, deterministic-but-unrelated path for every product, discarding the
source element's real IFC GlobalId. Converting IFC to IFCX silently broke
every external reference keyed on that GlobalId (a BCF topic, a diff, an
asset register). The TS exporter (packages/export/src/ifc5-exporter.ts)
already preferred GlobalId when building paths; the Rust port had dropped
that behavior.

path_for_id now reuses an entity's own GlobalId as its IFCX path when it has
one, falling back to the existing deterministic UUID-shaped path only for
entities with no GlobalId (e.g. the IfcProject root).

Split the growing test module into ifc5_tests.rs to stay under the 400-line
module-size ratchet (AGENTS.md), matching the schema_convert.rs /
schema_convert_tests.rs precedent.
